Allow hostname to be retained when selecting new network from network scanner
